Opening paragraph needs a little work.
The first sentence was worded weird so i made an adjustment, but I think the explanation of the "struggle" still doesn't sound right. I didn't bother putting my two cents into that though because I am not familiar with being a developer. Finally I felt that "But that's not all" reads better than "But that's not it".